Update solution for la-vector-projection

+++ study-plans/math-linear-algebra/la-vector-projection/la-vector-projection.py
@@ -4,8 +4,8 @@ def vector_projection(u, v):
     """
     Returns: float64 array, the projection of u onto v.
     """
-    u = np.asarray(u)
-    v = np.asarray(v)
+    u = np.asarray(u,dtype =float)
+    v = np.asarray(v,dtype = float)
     num = np.dot(v,u)
     den = np.dot(v,v)
     res = num/den * v
